requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
audeer==2.3.1
audformat==1.3.2
audinterface==1.3.1
audiofile==1.5.1
audioread==3.1.0
audmath==1.4.4
audresample==1.3.5
av==12.3.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
ctranslate2==4.6.1
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
defusedxml==0.7.1
demucs==4.0.1
denoiser @ git+https://github.com/facebookresearch/denoiser.git@8afd7c166699bb3c8b2d95b6dd706f71e1075df0
diffq==0.2.4
dora_search==0.1.12
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
faster-whisper==1.0.3
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fqdn==1.5.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.4
humanfriendly==10.0
hydra-colorlog==0.1.4
hydra-core==1.3.2
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
iso3166==2.1.1
iso639-lang==2.6.3
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lameenc==1.8.1
lark==1.3.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.45.1
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
musdb==0.4.3
museval==0.4.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.4.7
notebook_shim==0.2.4
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
openunmix==1.3.0
overrides==7.7.0
oyaml==1.0
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
panns-inference==0.1.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.7.0
pyarrow==22.0.0
pycparser==2.23
Pygments==2.19.2
pyloudnorm==0.1.1
pyparsing==3.2.5
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
stempeg==0.2.6
submitit==1.5.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchlibrosa==0.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
treetable==0.2.6
triton==3.1.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.9.0
widgetsnbextension==4.0.15
